pyproject.toml
[build-ssytem]
requires = ["wheel", "setuptools"]
build-backend="setuptools.build_meta"

[project]
name = "pygmailotp"
version = "0.1.0"
description = "gets emails from gmail and parses otps"
requires-python = ">=3.10"

[tools.setuptools.packages]
include = "pygmailotp"
exclude = "config"


[tool.poetry]
name = "pygmailotp"
version = "0.1.0"
description = "Bad code for doing bad things"
authors = ["FH"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.149.0"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
